pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agi"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9",
    "aiosqlite>=0.20",
    "litellm>=1.40",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-telegram-bot>=21.0",
    "pyautogui>=0.9",
    "mss>=9.0",
    "mcp>=1.0",
    "apscheduler>=3.10",
    "typer>=0.12",
    "pyyaml>=6.0",
    "sqlite-vec>=0.1",
    "pillow>=10.0",
    "aiofiles>=23.0",
    "httpx>=0.27",
    "discord.py>=2.3",
    "playwright>=1.40",
    "edge-tts>=6.1",
    "prompt_toolkit>=3.0",
]

[project.scripts]
agi = "agi.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/agi"]
